Many changes, still not working
So basicly it was a pain in the ass to re-think in react, because i know
i need redux in the future. So how do i make a small cklick-dummy if i
dont have the redux architecture?. Plus ... i dont think my webpack isnt
working, but how do i know? im just a trainee